Improve overloaded function call diagnostics (#1592)
From Solidity 0.6 onwards, when an overloaded function/event resolves to
multiple candidates, this is an error. In earlier versions, the first
result is used. So, use the `pragma solidity` version to decide whether
to error or not.

Fixes: https://github.com/hyperledger/solang/issues/1534
Fixes: https://github.com/hyperledger/solang/issues/707

+++ src/sema/ast.rs
@@ -742,11 +742,11 @@ pub enum VersionReq {
     },
 }
 
-#[derive(Debug)]
+#[derive(Debug, Clone, PartialEq)]
 pub struct Version {
-    pub major: u64,
-    pub minor: Option<u64>,
-    pub patch: Option<u64>,
+    pub major: u32,
+    pub minor: Option<u32>,
+    pub patch: Option<u32>,
 }
 
 impl fmt::Display for Version {
